requirements.txt
Django==2.2.24
djangorestframework==3.12.4
celery==4.4.7
